controller: add SppPcap class
it is easy to send a request to spp_pcap from 'spp.py'
without any effort. This update is to add SppPcap class
behaviour as a client for spp-ctl. An instance of the class is
intended to be used from do_pcap() and complete_pcap() methods of
Shell class in 'spp.py'.

Signed-off-by: Hideyuki Yamashita <yamashita.hideyuki@po.ntt-tx.co.jp>
Signed-off-by: Naoki Takada <takada.naoki@lab.ntt.co.jp>