controller: add pcap command to SPP controller
To manage spp_pcap from 'pcap.py', add 'pcap' command to Shell class.
Spp_pcap might have several instances as similar to spp_pcap,
and deciding which of ones is also similar to. 'pcap' command consists
of a indicator and actual command. Here is an example.

  spp > pcap 3; start

In this example, indicator 'pcap 3;' is before spp_pcap's command
'capture start'. The number in indicator is a secondary ID actually,
so you cannot assign the same ID of others.

You can refer the usage of 'pcap' command with 'help' command.

  spp > help pcap

Signed-off-by: Hideyuki Yamashita <yamashita.hideyuki@po.ntt-tx.co.jp>
Signed-off-by: Naoki Takada <takada.naoki@lab.ntt.co.jp>