spp_vf: fix bug of fixing slow command reply
This update is to fix incorrect fixing of slow command reply.

To reduce time for making command reply, replace sleep(1) to usleep(100)
for, and add another sleep(1) to avoid sending too much requests if sock
is closed and retry in previous patches. However, response is still slow
because the place of added sleep(1) is incorrect.

Fixes: ab5183201c02 ("spp_vf: add sleep after disconnection detection")
Fixes: 5eea2dfa5420 ("fix slow response of command")

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>