spp_vf: remove while loop in classifier_mac
So far, spp_classifier_mac_do() has while loop and it does not return
until the classifier component is requested to stop and thus the worker
thread can not execute packet processing for forwarder/merger.
To introduce core-sharing, this patch removes the while loop
to make classifier not occupy assigned core.

Signed-off-by: Hideyuki Yamashita <yamashita.hideyuki@po.ntt-tx.co.jp>
Signed-off-by: Naoki Takada <takada.naoki@lab.ntt.co.jp>