spp_mirror: remove ifdef switch
There are some files used in common for spp_vf and spp_mirror
but actually different slightly.
For some logic, so far ifdef switch is used.
However this may cause complexity in the code and thus create
difficulity when adding new secondary process.
To improve this, this patch moves such files under mirror directory
to remove dependency.

To delete the handling of that is unnecessary for deletion and
'spp_mirror' ifdef declaration of a component copying the following
files than vf/common subordinates.

Signed-off-by: Hideyuki Yamashita <yamashita.hideyuki@po.ntt-tx.co.jp>
Signed-off-by: Naoki Takada <takada.naoki@lab.ntt.co.jp>